cont

+++ utils/wav.h
@@ -1,2 +1,2 @@
 
-int read_wav_file(char *filename, float **data, int *num_chan, int *num_items, int *sample_rate);
+int read_wav_file(char *filename, double **data, int *num_chan, int *num_items, int *sample_rate);
